Backend/exun23/requirements.txt
asgiref==3.7.2
boto3==1.28.73
botocore==1.31.73
Django==4.2.6
django-cors-headers==4.3.0
django-storages==1.14.2
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
jmespath==1.0.1
packaging==23.2
Pillow==10.1.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
